scrapper-microservice/requirements.txt
gitpython==3.1.36